user rank in stats added
