+++ src/components/stats.js
@@ -3,6 +3,7 @@ import Chart from "chart.js";
 import ChartDataLabels from "chartjs-plugin-datalabels";
 import {filmDuration} from "../utils/render.js";
 import {getMaxValueKeyFromObject} from "../utils/common.js";
+import {userLevel} from "../utils/filter.js";
 
 const createPeriodsMarkup = (period, isActive) => {
   const periodInTags = period.toLowerCase().replace(` `, `-`);
@@ -20,12 +21,14 @@ const createStatisticsTemplate = (films, topGenre, Periods) => {
     return a + b.duration;
   }, 0));
 
+  const userRank = userLevel(films);
+
   return (
     `<section class="statistic">
     <p class="statistic__rank">
       Your rank
       <img class="statistic__img" src="images/bitmap@2x.png" alt="Avatar" width="35" height="35">
-      <span class="statistic__rank-label">Sci-Fighter</span>
+      <span class="statistic__rank-label">${userRank}</span>
     </p>
 
     <form action="https://echo.htmlacademy.ru/" method="get" class="statistic__filters">
