+++ src/utils/filter.js
@@ -1,5 +1,12 @@
 import {FilterType} from "../const.js";
 
+const UserLevels = {
+  0: ``,
+  10: `Novice`,
+  20: `Fan`,
+  21: `Movie Buff`,
+};
+
 const getWatchlistFilms = (films) => {
   return films.filter((film) => film.addToWatchList);
 };
@@ -27,4 +34,18 @@ const getFilmsByFilter = (films, filterType) => {
   return films;
 };
 
-export {getFilmsByFilter};
+const userLevel = (films) => {
+  let userRank = UserLevels[`21`];
+  const watchedFilmsCount = films.filter((film) => film.alreadyWatched === true).length;
+
+  for (let i = 0; i < Object.keys(UserLevels).length; i++) {
+    if (watchedFilmsCount <= parseInt(Object.keys(UserLevels)[i], 10)) {
+      userRank = Object.values(UserLevels)[i];
+      break;
+    }
+  }
+
+  return userRank;
+};
+
+export {getFilmsByFilter, userLevel};
